iex-stream/cloud_functions/iex_api/requirements.txt
google-cloud==0.34.0
google-cloud-pubsub==1.7.0
